+++ compute_modules/context/__init__.py
@@ -13,12 +13,10 @@
 #  limitations under the License.
 
 
-from ._version import __version__ as __version__
-from .function_registry.function_registry import add_function, add_functions
-from .startup import start_compute_module
+from .context import get_extra_context_parameters
+from .types import QueryContext
 
 __all__ = [
-    "add_function",
-    "add_functions",
-    "start_compute_module",
+    "get_extra_context_parameters",
+    "QueryContext",
 ]
